AGS: Common: moved iagsstream to util dir (api dir proved to be excessive)
From upstream 2de0701f47cb189b51aa429cd8ef4812b6dae9ef
