+++ engines/ags/shared/util/iags_stream.h
